add mp to template

+++ fft.cpp
@@ -14,7 +14,7 @@ void fft(vector<C> &a) {
     root[1] = croot[1] = 1.0;
 
     for (int k = 2; k < n; k *= 2) {
-        complex<long double> x = polar(1.0L, acos(-1.0L) / k);
+        complex<long double> x = polar(1.0L, acosl(-1.0L) / k);
 
         for (int i = k; i < 2 * k; i++) {
             root[i] = croot[i] = i & 1 ? croot[i / 2] * x : croot[i / 2];
